Cleaned modal-slideshow.js comments and improved naming for modal objects.
Modal objects were initially named in accordance with the HTML code provided by Treehouse, but their naming scheme was not optimal, and my wrapper to group the modal card and the button container added another element.

So I changed things around in all the relevant documents and tested the modal slideshow to make sure it still worked.